MAT-4259 create measure resource from madie measure
